Move requirePaymentConfirmed to app-request-handler
