+++ src/api/server.ts
@@ -1,14 +1,11 @@
 import cookieParser from 'cookie-parser';
 import cors from 'cors';
-import express, { RequestHandler } from 'express';
+import express from 'express';
 import { readFileSync } from 'fs';
 import helmet from 'helmet';
 import http from 'http';
 import https from 'https';
 import { apiBasePath, ApiPath } from '../domain/api-path';
-import { isAccountNotFound } from '../domain/account';
-import { PlanId } from '../domain/plan';
-import { loadAccount } from '../domain/account-storage';
 import { startCronJob } from '../shared/cron-utils';
 import { getErrorMessage } from '../shared/lang';
 import { logHeartbeat, makeCustomLoggers } from '../shared/logging';
@@ -22,7 +19,7 @@ import {
   requestAccountPasswordChange,
   requestAccountPlanChange,
 } from './account';
-import { makeAppRequestHandler } from './app-request-handler';
+import { makeAppRequestHandler, requirePaymentConfirmed } from './app-request-handler';
 import { authentication } from './authentication';
 import { deauthentication } from './deauthentication';
 import { deliveryReports } from './delivery-reports';
@@ -43,7 +40,6 @@ import { showSampleEmail, showSampleEmailPublic } from './feeds/show-sample-emai
 import { initApp } from './init-app';
 import { confirmPasswordReset, requestPasswordReset } from './password-reset';
 import { registration, registrationConfirmation } from './registration';
-import { checkSession } from './session';
 import { makeExpressSession } from './session';
 import { sessionTest } from './session-test';
 import {
@@ -56,28 +52,6 @@ import {
 import { subscription } from './subscription';
 import { subscriptionConfirmation } from './subscription-confirmation';
 import { unsubscription } from './unsubscription';
-import { isErr } from '../shared/lang';
-import { App } from './init-app';
-
-function requirePaymentConfirmed(app: App): RequestHandler {
-  return (req, res, next) => {
-    const session = checkSession(req.session);
-
-    if (session.kind !== 'AuthenticatedSession') {
-      next();
-      return;
-    }
-
-    const account = loadAccount(app.storage, session.accountId);
-
-    if (!isErr(account) && !isAccountNotFound(account) && account.planId === PlanId.PendingPayment) {
-      res.status(402).json({ kind: 'AppError', message: 'Payment confirmation pending' });
-      return;
-    }
-
-    next();
-  };
-}
 
 async function main() {
   const { logInfo, logWarning } = makeCustomLoggers({ module: 'api-server' });
